backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# API clients
httpx==0.26.0
elevenlabs==0.2.27
anthropic==0.18.1
openai==1.12.0

# Audio/Video processing
pydub==0.25.1
numpy==1.26.3
pillow==10.2.0
opencv-python-headless==4.9.0.80

# Database and storage
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Utilities
python-dotenv==1.0.1
pydantic==2.6.0
pydantic-settings==2.1.0
aiofiles==23.2.1

# Background tasks
celery==5.3.6
redis==5.0.1
